backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.9.2
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
certifi>=2024.0.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx>=0.27.0
openai>=2.0.0
